src/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
boto3==1.33.0
pytest==7.4.3
pandas==2.1.3
pyarrow==13.0.0
PyJWT==2.8.1
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
azure-identity==1.14.0
azure-storage-blob==12.18.3
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.0.1
python-dotenv==1.0.0
httpx==0.25.2
salesforce-bulk==2.3.0
simple-salesforce==1.12.5
snowflake-connector-python==3.5.0
pyspark==3.5.0
kafka-python==2.0.2
cx-Oracle==8.3.0
sqlalchemy==1.4.48
mangum==0.17.0